Broke out workspace controls into subunits. Conditionally display grid-related controls.
